updated to use spiffworkflow-connector-command w/ burnettk
