+++ .tool-versions
@@ -1,0 +1,1 @@
+python 3.11.0
