+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "connector-smtp"
-version = "0.1.0"
+version = "1.0.0"
 description = "Make SMTP Requests available to SpiffWorkflow Service Tasks"
 authors = ["Jon Herron <jon.herron@yahoo.com>"]
 readme = "README.md"
@@ -8,8 +8,60 @@ packages = [{include = "connector_smtp", from = "src" }]
 
 [tool.poetry.dependencies]
 python = "^3.9"
+spiffworkflow-connector-command = {git = "https://github.com/sartography/spiffworkflow-connector-command.git", rev = "main"}
+
+[tool.poetry.group.dev.dependencies]
+mypy = "^1.6.0"
+ruff = "^0.0.292"
+pytest = "^7.4.2"
 
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.ruff]
+select = [
+  "B",   # flake8-bugbear
+  "C",   # mccabe
+  "E",   # pycodestyle error
+  # "ERA",  # eradicate
+  "F",   # pyflakes
+  "N",   # pep8-naming
+  "PL",  # pylint
+  "S",   # flake8-bandit
+  "UP",  # pyupgrade
+  "W",   # pycodestyle warning
+  "I001" # isort
+]
+
+ignore = [
+  "C901",  # "complexity" category
+  "PLR",  # "refactoring" category has "too many lines in method" type stuff
+  "PLE1205"  # saw this Too many arguments for `logging` format string give a false positive once
+]
+
+line-length = 130
+
+# target python 3.10
+target-version = "py310"
+
+exclude = [
+  "migrations"
+]
+
+[tool.ruff.per-file-ignores]
+"migrations/versions/*.py" = ["E501"]
+"tests/**/*.py" = ["PLR2004", "S101"]  # PLR2004 is about magic vars, S101 allows assert
+
+[tool.ruff.isort]
+force-single-line = true
+
+[tool.mypy]
+strict = true
+disallow_any_generics = false
+warn_unreachable = true
+pretty = true
+show_column_numbers = true
+show_error_codes = true
+show_error_context = true
